unit test fix
