+++ test/unit/general.controller.test.ts
@@ -5,9 +5,58 @@ import tdeiCoreService from "../../src/service/tdei-core-service";
 import jobService from "../../src/service/job-service";
 import { InputException } from "../../src/exceptions/http/http-exceptions";
 import { getMockFileEntity } from "../common/mock-utils";
+import { DatasetDTO } from "../../src/model/dataset-dto";
 
 // group test using describe
 describe("General Controller Test", () => {
+
+    describe("Get Dataset list", () => {
+        describe("Functional", () => {
+            test("When requested with empty search criteria, Expect to return Dataset list", async () => {
+                //Arrange
+                const req = getMockReq();
+                const { res, next } = getMockRes();
+                const list: DatasetDTO[] = [<DatasetDTO>{}]
+                const getDatasetsSpy = jest
+                    .spyOn(tdeiCoreService, "getDatasets")
+                    .mockResolvedValueOnce(list);
+                //Act
+                await generalController.getDatasetList(req, res, next);
+                //Assert
+                expect(getDatasetsSpy).toHaveBeenCalledTimes(1);
+                expect(res.status).toHaveBeenCalledWith(200);
+                expect(res.send).toBeCalledWith(list);
+            });
+
+            test("When requested with bad collection_date input, Expect to return HTTP status 400", async () => {
+                //Arrange
+                const req = getMockReq({ body: { collection_date: "2023" } });
+                const { res, next } = getMockRes();
+                jest
+                    .spyOn(tdeiCoreService, "getDatasets")
+                    .mockRejectedValueOnce(new InputException("Invalid date provided."));
+                //Act
+                await generalController.getDatasetList(req, res, next);
+                //Assert
+                expect(res.status).toHaveBeenCalledWith(400);
+                expect(next).toHaveBeenCalled();
+            });
+
+            test("When unknown or database exception occured while processing request, Expect to return HTTP status 500", async () => {
+                //Arrange
+                const req = getMockReq({ body: { collection_date: "2023" } });
+                const { res, next } = getMockRes();
+                jest
+                    .spyOn(tdeiCoreService, "getDatasets")
+                    .mockRejectedValueOnce(new Error("unknown error"));
+                //Act
+                await generalController.getDatasetList(req, res, next);
+                //Assert
+                expect(res.status).toHaveBeenCalledWith(500);
+                expect(next).toHaveBeenCalled();
+            });
+        });
+    });
 
     describe('Invalidate the record', () => {
